refactor: move elementtemplate class and field classes into model dir
