--- connect/connect-c7/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/Main.java
@@ -1,13 +1,0 @@
-package io.miragon.miranum.connect.adapter.in.c7.elementtemplates;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Import;
-
-@SpringBootApplication
-@Import(Test.class)
-public class Main {
-    public static void main(String[] args) {
-        SpringApplication.run(Main.class, args);
-    }
-}
