--- connect/connect-c7/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/Test.java
@@ -1,25 +1,0 @@
-package io.miragon.miranum.connect.adapter.in.c7.elementtemplates;
-
-import io.miragon.miranum.connect.elementtemplate.api.BPMNElementType;
-import io.miragon.miranum.connect.elementtemplate.api.GenerateElementTemplate;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-public class Test {
-    @GenerateElementTemplate(name = "Test", id = "test", appliesTo = {BPMNElementType.BPMN_SERVICE_TASK}, version = "1.0.0")
-    public TestOutput test(TestInput input) {
-        return new TestOutput(input.getName(), input.getId());
-    }
-}
-
-@Getter
-class TestInput {
-    private String name;
-    private String id;
-}
-
-@AllArgsConstructor
-class TestOutput {
-    private String name;
-    private String id;
-}
