+++ connect/connect-c7/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/model/Binding.java
@@ -1,0 +1,16 @@
+package io.miragon.miranum.connect.adapter.in.c7.elementtemplates.model;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public
+class Binding {
+    private String type;
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String name;
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private String source;
+}
