+++ connect/connect-c7/element-templates-c7/src/main/java/io/miragon/miranum/connect/adapter/in/c7/elementtemplates/model/Constraints.java
@@ -1,0 +1,10 @@
+package io.miragon.miranum.connect.adapter.in.c7.elementtemplates.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+class Constraints {
+    private boolean notEmpty;
+}
